feat!: return nil to automatically stop iteration - rather than -1
Previously, -1 was returned to signal the end of parsing, just like
POSIX getopt. However, an iterator was being made use of to make usage
more user-friendly. This commit goes all the way and instead of
returning -1, nil is returned as usual to stop Lua's `for` iterator
automatically.
However, in the absence of returning -1, nothing else -- including
optind -- is returned. This means that the user won't know where to
resume parsing if there are argv elements left. The solution adopted
here is for the getopt functions to take an (optional) `leftovers`
argument - a table that they will populate with all the remaining arg
elements that were not consumed. The user can then easily iterate over
these and do whatever they like with the remaining argv elements placed
in this `leftovers` table.